Fixed bug in calculation of link shaping curve.
Now the path a flow traverses in the network is a property of the flow.
When setting the path the MFS of the flow is set as the max. packet
size of all links within the path if the MFS is the maximum.

Else the sequential adding of the flows at the server graph will
calculate wrong shaping curves depending on the order in which the flows
are added.