Update of CBS queues and TFA calculation.
The calculation of the servers CBS queues was moved to new funciton.
Needs to be called after adding all flows and their paths to the server
graph. Then all CBS queues for all flows are calculated once starting
with the highest to lowest priority.

The server graph can now also be set with various support for shaping:
* no shaping
* link shaping only
* CBS shaping only
* link and CBS shaping

In addition TFA calculation was updated as well and uses now the Num
datatype instead of double to better match the original non-CBS
implementation.